+++ src/meltano/api/workers/__init__.py
@@ -1,4 +1,3 @@
 from .meltano_compiler_worker import MeltanoCompilerWorker
-from .airflow_worker import AirflowWorker
 from .ui_available_worker import UIAvailableWorker
 from .api_worker import APIWorker