+++ src/meltano/__init__.py
@@ -1,2 +1,2 @@
 # Managed by bumpversion
-__version__ = "1.37.1"
+__version__ = "1.38.0"